requirements.txt
Django>=5.0
psycopg2-binary>=2.9
djangorestframework>=3.15
djangorestframework-simplejwt>=5.3
PyYAML>=6.0
python-dateutil>=2.9